// src/types/common/common.types.ts
export interface Language {
  code: string;
  name: string;
}

export interface TranslationResponseType {
  translatedText: string;
  source_lang: string;
}

export interface DetectionWithTranslation {
  mid: string;
  name: string;
  translatedName: string;
  score: number;
  description?: string;
  boundingPoly: {
    normalizedVertices: [
      { x: number; y: number },
      { x: number; y: number },
      { x: number; y: number },
      { x: number; y: number }
    ];
  };
  translations: {
    language: string;
    translation: string;
    description: string;
  }[];
  descriptions: { language: string; description: string; }[];
}

export interface ObjectDetectionFromImageResponseType {
  image_name: string;
  detections: DetectionWithTranslation[]
}